# Conflicts:
#	apps/api/src/comment.controller.ts
#	apps/api/src/comment/comment.module.ts
#	libs/shared/src/modules/postgresdb/postgresdb.module.ts

+++ apps/api/src/comment/comment.service.ts
@@ -1,20 +1,30 @@
-import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
+import { HttpException, HttpStatus, Injectable, UnauthorizedException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/sequelize";
 import { CommentFilm } from "@app/shared/models/comment.model";
 import { CreatCommentFilmDto } from "../../../../libs/shared/src/dtos/comment-dto/creatCommentFilm.dto";
 import { Op } from "sequelize";
+import { JwtService } from "@nestjs/jwt";
+import { CreateProfileDto } from "@app/shared/dtos/auth-dto/create-profile.dto";
+import { Profile } from "@app/shared/models/profile.model";
+import { UserRoles } from "@app/shared/models/user-role.model";
+import * as process from "process";
 
 @Injectable()
 export class CommentService {
 
-  constructor(@InjectModel(CommentFilm) private commentFilmRepository: typeof CommentFilm) {
+  constructor(@InjectModel(CommentFilm) private commentFilmRepository: typeof CommentFilm,
+              private jwtService: JwtService,
+              @InjectModel(Profile) private profileRepository: typeof Profile,
+              @InjectModel(UserRoles) private userRoleRepository: typeof UserRoles
+  ) {
   }
 
   async getCommentsByFilmId(id: number) {
 
     const comments = await this.commentFilmRepository.findAll({ where: { film_id: id }, raw: true });
-    console.log("comments "  + comments.length );
+    console.log("comments " + comments.length);
     if (comments.length === 0) throw new HttpException("Нет комментариев к этому фильму", HttpStatus.BAD_REQUEST);
+
     function buildHierarchy(arry) {
 
       var roots = [], children = {};
@@ -57,10 +67,20 @@ export class CommentService {
     }
   }
 
-  async deleteCommentFilm(id: number) {
+  async deleteCommentFilm(req: Request, id: number) {
+    const authHeader = req.headers["authorization"];
+    const token = authHeader.split(" ")[1];
+    const userAuth = this.jwtService.verify<CreateProfileDto>(token, {
+      secret: process.env.PRIVATE_KEY
+    });
     const comment = await this.commentFilmRepository.findOne({ where: { id: id } });
     if (comment) {
-      await this.commentFilmRepository.destroy({ where:  {[Op.or] : [{id: comment.id}, {parentCommentId: comment.id} ]}  }  );
+      const profileRepository = await this.profileRepository.findOne({ where: { email: userAuth.email } });
+      const roleIdUserAuth = await this.userRoleRepository.findOne({ where: { profileId: profileRepository.id } });
+      if (roleIdUserAuth.roleId != 1 && comment.profileId !== profileRepository.id) {
+        throw new UnauthorizedException({ message: "Комментарий не удален, пользователь не админ, либо он не автор комментария" });
+      }
+      await this.commentFilmRepository.destroy({ where: { [Op.or]: [{ id: comment.id }, { parentCommentId: comment.id }] } });
       return comment.id;
     } else {
       throw new HttpException("Такого комментария не существует", HttpStatus.BAD_REQUEST);
